Refactor processRow
Moves processRow to utils so it can be reused by the grid view
